Migrations
There are no further differences between production and the current master code, no other migrations are needed
Once we update live we need to execute a SQL to register the initial migration as "executed", from then on we
can use migrations normally.

The adjustment file is in doc/db for historic purposes.